+++ lib/suitup.dart
@@ -1,5 +1,6 @@
 library suitup;
 
+export 'src/suitup_cancellable.dart';
 export 'src/suitup_color_abstract.dart';
 export 'src/suitup_controller.dart';
 export 'src/suitup_extensions.dart';